\name{anes04}
\alias{anes04}
\docType{data}
\title{
Demographic Data From 2004 American National Election Studies (ANES)
}
\description{
A dataset containing demographic data from the 2004 American National Election Studies.  The data include 5 variables: "female" (A Logical Variable Indicating Sex), "age" (Numerically Coded, Ranging From 18 to a Topcode of 90), "educats" (5 Education Categories corresponding to 1-Less than A High School Degree, 2-High School Gradutate, 3-Some College, 4-College Graduate, 5-Post College Education), "racecats" (6 Racial Categories), and "married" (A Logical Variable Indicating the Respondent's Marital Status, with one point of missing data).  Dataset is designed show how production of survey weights works in practice.
}
\usage{data(anes04)}
\format{
  The format is:
 chr "anes04"
}
\source{
http://www.electionstudies.org
}
\keyword{datasets}
