\name{generaldesigneffect}
\alias{generaldesigneffect}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Calculates a general design effect given weights for a dataset.
}
\description{
Calculates a general design effect given weights for a dataset.
}
\usage{
generaldesigneffect(weightvec)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{weightvec}{
Vector of weights.
}
}